feat(ngParser): .child method + 2nd argument for custom options
The .child method is for creating new instances of ngParser
in which the default options can be changed.
This is because expressions parsed with custom options
can't be cached. So the way around this is to create a new
instance with different default options and a cache on it's own